Preparing RestoreSession init container for OnlineKind restores
